Made the ConfigChangeListener its own class for reuse elsewhere
